+++ kibot/kicad/v5_sch.py
@@ -906,6 +906,8 @@ class SchematicComponent(object):
         self.has_pcb_info = False
         self.qty = 1
         self.annotation_error = False
+        # Position offset i.e. from the rotation filter
+        self.pos_offset_x = self.pos_offset_y = None
         # KiCad 5 PCB flags (mutually exclusive)
         self.smd = False
         self.virtual = False